fix: show contribution banner on wiki space details page
Move ContributionBanner from child components (WikiDocumentPanel,
DraftContributionPanel) to the parent SpaceDetails layout so it
remains visible on all routes including the space welcome page.
